Move most Dialog logic to its component
This was tricky… the use of `useImperativeHandle` was the only way we could dynamically pass `confirmCallback` to `<Dialog>`. Other attempts (like using props) failed, probably because the `successCallback` must be dynamic. Who knows… At the end, triggering `showModal()` from the parent with functions as parameters is working. Likely not a nice way to do it, but might be okay-ish for now.